requirements.txt
python-fasthtml 
pokebase
